Add check for ENOENT to unlink

+++ angr/procedures/linux_kernel/unlink.py
@@ -15,6 +15,11 @@ class unlink(angr.SimProcedure): #pylint:disable=W0622
         str_expr = self.state.memory.load(path_addr, p_strlen.max_null_index, endness='Iend_BE')
         str_val = self.state.solver.eval(str_expr, cast_to=bytes)
 
+        # Check if entity exists before attempting to unlink
+        if not self.state.fs.get(str_val):
+            self.state.libc.ret_errno('ENOENT')
+            return -1
+
         if self.state.fs.delete(str_val):
             return 0
         else:
